+++ src/lib/validation/at-rules/when.ts
@@ -3,6 +3,7 @@ import type {ValidationSyntaxResult} from "../../../@types/validation";
 import {EnumToken, ValidationLevel} from "../../ast";
 import {consumeWhitespace, splitTokenList} from "../utils";
 import {validateMediaCondition, validateMediaFeature} from "./media";
+import {validateSupportCondition} from "./supports";
 
 export function validateAtRuleWhen(atRule: AstAtRule, options: ValidationOptions, root?: AstNode): ValidationSyntaxResult {
 
@@ -75,7 +76,7 @@ export function validateAtRuleWhenQueryList(tokenList: Token[], atRule: AstAtRul
 
         while (split.length > 0) {
 
-            if (split[0].typ != EnumToken.FunctionTokenType || !['media', 'supports'].includes((split[0] as FunctionToken).val)) {
+            if (split[0].typ != EnumToken.FunctionTokenType || !['media', 'supports', 'font-tech', 'font-format'].includes((split[0] as FunctionToken).val)) {
 
                 result = {
                     valid: ValidationLevel.Drop,
@@ -110,10 +111,10 @@ export function validateAtRuleWhenQueryList(tokenList: Token[], atRule: AstAtRul
                     break;
                 }
 
-            } else if (split[0].val == 'supports') {
+            } else if (['supports', 'font-tech', 'font-format'].includes(split[0].val)  ) {
 
                 // result = valida
-                if (!validateMediaCondition(split[0], atRule)) {
+                if (!validateSupportCondition(atRule, split[0])) {
 
                     result = {
                         valid: ValidationLevel.Drop,